+++ pytorch_lightning/loops/base.py
@@ -17,15 +17,12 @@ from collections import OrderedDict
 from typing import Any, Dict, Optional
 
 from deprecate import void
+from torch.nn.modules.module import _IncompatibleKeys
 
 import pytorch_lightning as pl
-from pytorch_lightning.trainer.progress import BaseProgress, ProgressDict, Tracker
+from pytorch_lightning.trainer.progress import BaseProgress, Tracker
 from pytorch_lightning.utilities.apply_func import apply_to_collection
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.model_helpers import is_overridden
-from pytorch_lightning.utilities.warnings import WarningCache
-
-warning_cache = WarningCache()
 
 
 class Loop(ABC):
@@ -52,88 +49,35 @@ class Loop(ABC):
 
     def __init__(self) -> None:
         self.iteration_count: int = 0
-        self.trainer: Optional['pl.Trainer'] = None
-        self._loops = OrderedDict()
-        self._progress = OrderedDict()
-        self._has_parent: bool = False
-        self.__parent_loop: Optional['Loop'] = None
+        self._trainer: Optional['pl.Trainer'] = None
         self.restarting = False
-
-    @property
-    def has_parent(self) -> Optional[bool]:
-        """Whether this loop has been attached to another loop"""
-        return self._has_parent
-
-    @property
-    def has_children(self) -> bool:
-        """Whether this loop has any children"""
-        loops = self.__dict__.get('_loops')
-        return len(loops) > 0
-
-    @property
-    def is_leaf(self) -> bool:
-        """This loop is a leaf if it doesn't possess any loops."""
-        return not self.has_children
 
     @property
     def loop_progress(self) -> Dict[str, Any]:
         """Return the progress for the current loop and children loop."""
         progress = {}
-        for n, p in self.__dict__.get('_progress').items():
-            progress[n] = p
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                progress[k] = v
+            elif isinstance(v, Loop):
+                progress[k] = v.loop_progress
+        return progress
 
-        loops = self.__dict__.get('_loops')
+    @property
+    def trainer(self) -> Optional['pl.Trainer']:
+        return self._trainer
 
-        if loops is not None:
-            for name, loop in loops.items():
-                progress[name] = ProgressDict(**loop.loop_progress)
-        return ProgressDict(**progress)
-
-    def __setattr__(self, name: str, value: Any) -> None:
-        if isinstance(value, pl.Trainer) or (name == "restarting" and value is True):
-            object.__setattr__(self, name, value)
-            for loop in self._loops.values():
-                loop.__setattr__(name, value)
-        elif isinstance(value, Loop):
-            if name == "_Loop__parent_loop":
-                object.__setattr__(self, name, value)
-                return
-            if getattr(self, "__children__loops__", None) is not None and name not in self.__children__loops__:
-                raise MisconfigurationException(
-                    f"The current loop accept only {self.__children__loops__} as children attribute names. Found {name}"
-                )
-            if value._has_parent:
-                raise MisconfigurationException(f"This provided loop {value} already has a parent. ")
-            self._loops[name] = value
-            value._has_parent = True
-            value.__parent_loop = self
-        elif isinstance(value, BaseProgress):
-            self._progress[name] = value
-        else:
-            object.__setattr__(self, name, value)
-
-    def __getattr__(self, name) -> Any:
-        loops = self.__dict__.get('_loops')
-
-        if loops is not None and name in loops:
-            return loops[name]
-
-        progress = self.__dict__.get('_progress')
-
-        if progress is not None and name in progress:
-            return progress[name]
-
-        return object.__getattribute__(self, name)
-
-    def __delattr__(self, name) -> None:
-        if name in self._loops:
-            self._loops[name]._has_parent = False
-            self._loops[name]._Loop__parent_loop = None
-            del self._loops[name]
-        elif name in self._progress:
-            del self._progress[name]
-        else:
-            object.__delattr__(self, name)
+    @trainer.setter
+    def trainer(self, trainer: 'pl.Trainer'):
+        """Connect the Trainer to itself and all sub-children loops"""
+        if not isinstance(trainer, pl.Trainer):
+            raise MisconfigurationException(
+                f"Loop {self.__class__.__name__} should be connected to a `Trainer`, found: {trainer}."
+            )
+        self._trainer = trainer
+        for v in self.__dict__.values():
+            if isinstance(v, Loop):
+                v.trainer = trainer
 
     @property
     @abstractmethod
@@ -148,10 +92,6 @@ class Loop(ABC):
     def connect(self, trainer: 'pl.Trainer', *args: Any, **kwargs: Any) -> None:
         """Connects Loop with all the necessary things like connectors and accelerators."""
         # TODO(@justusschock): Make the trainer a weakref/proxy
-        if not isinstance(trainer, pl.Trainer):
-            raise MisconfigurationException(
-                f"Loop {self.__class__.__name__} should be connected to a `Trainer`, found: {trainer}."
-            )
         self.trainer = trainer
 
     def on_skip(self) -> Optional[Any]:
@@ -178,14 +118,7 @@ class Loop(ABC):
         if self.skip:
             return self.on_skip()
 
-        if self.restarting:
-            if not is_overridden("restore", self, Loop):
-                warning_cache.warn(f"{self.__class__.__name__} Loop doesn't override the restore function.")
-            is_restore_finished = self.restore()
-            if is_restore_finished is None or is_restore_finished:
-                self.restarting = False
-        else:
-            self.reset()
+        self.reset()
 
         self.on_run_start(*args, **kwargs)
 
@@ -200,14 +133,6 @@ class Loop(ABC):
 
         output = self.on_run_end()
         return output
-
-    def restore(self) -> Optional[bool]:
-        """
-        Restore the internal state of the loop the beginning of run if restarting is ``True``.
-
-        Returns:
-            is_restore_finished: Whether the restoration actually finished.
-        """
 
     @abstractmethod
     def reset(self) -> None:
@@ -240,62 +165,81 @@ class Loop(ABC):
     def teardown(self) -> None:
         """Use to release memory etc."""
 
-    def state_dict(self) -> Dict:
-        """Current Loop state"""
+    def on_save_checkpoint(self) -> Dict:
+        """
+        Called when saving a model checkpoint, use to persist loop state.
+
+        Returns:
+            The current loop state.
+        """
         return {}
 
-    def load_state_dict(self, state_dict: Dict) -> None:
-        """Reload Loop state"""
+    def on_load_checkpoint(self, state_dict: Dict):
+        """Called when loading a model checkpoint, use to reload loop state."""
 
-    def get_state_dict(self, destination: Optional[OrderedDict] = None, prefix: Optional[str] = '') -> OrderedDict:
+    def state_dict(self, destination: Optional[OrderedDict] = None, prefix: Optional[str] = '') -> Dict:
         if destination is None:
             destination = OrderedDict()
 
-        destination[prefix + "state_dict"] = self.state_dict()
+        destination[prefix + "state_dict"] = self.on_save_checkpoint()
 
-        for name, progress in self._progress.items():
-            destination[prefix + name] = progress.state_dict()
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                destination[prefix + k] = v.state_dict()
+            elif isinstance(v, Loop):
+                v.state_dict(destination, prefix + k + '.')
 
-        for name, loop in self._loops.items():
-            loop.get_state_dict(destination, prefix + name + '.')
         return destination
 
-    def get_progress_state_dict(
-        self, destination: Optional[OrderedDict] = None, prefix: Optional[str] = ''
-    ) -> OrderedDict:
-        if destination is None:
-            destination = OrderedDict()
+    def _load_from_state_dict(self, state_dict, prefix, restart_progress, missing_keys, unexpected_keys, error_msgs):
+        for k, v in self.__dict__.items():
+            if isinstance(v, BaseProgress):
+                v.load_state_dict(state_dict[prefix + k])
 
-        for name, progress in self._progress.items():
-            destination[prefix + name] = progress.state_dict()
+                if restart_progress:
 
-        for name, loop in self._loops.items():
-            loop.get_progress_state_dict(destination, prefix + name + '.')
-        return destination
+                    def restart(v: Tracker):
+                        v.reset_on_restart()
 
-    def _load_from_state_dict(self, state_dict, prefix, apply_restart):
-        # reload progress first as it might be used for ``load_state_dict``.
-        for name, progress in self._progress.items():
-            progress.load_state_dict(state_dict[prefix + name])
-            if apply_restart:
+                    apply_to_collection(v, Tracker, restart)
 
-                def restart(v: Tracker):
-                    v.reset_on_restart()
+        self.on_load_checkpoint(state_dict[prefix + "state_dict"])
 
-                apply_to_collection(progress, Tracker, restart)
-
-        self.load_state_dict(state_dict[prefix + "state_dict"])
-
-    def _load_state_dict(self, state_dict: Dict, apply_restart: bool = True):
+    def load_state_dict(self, state_dict: Dict, restart_progress: bool = True, strict: bool = True):
+        """
+        This function is highly inspired from ``PyTorch nn.Module``.
+        """
+        missing_keys = []
+        unexpected_keys = []
+        error_msgs = []
 
         state_dict = state_dict.copy()
 
         def load(loop, prefix=''):
-            loop._load_from_state_dict(state_dict, prefix, apply_restart)
+            if loop.restarting:
+                return
+            loop._load_from_state_dict(state_dict, prefix, restart_progress, missing_keys, unexpected_keys, error_msgs)
             loop.restarting = True
-            for name, loop_children in loop._loops.items():
-                if loop_children is not None:
-                    load(loop_children, prefix + name + '.')
+            for k, v in self.__dict__.items():
+                if isinstance(v, Loop):
+                    load(v, prefix + k + '.')
 
         load(self)
-        load = None  # break load->load reference cycle
+
+        if strict:
+            if len(unexpected_keys) > 0:
+                error_msgs.insert(
+                    0, 'Unexpected key(s) in state_dict: {}. '.format(
+                        ', '.join('"{}"'.format(k) for k in unexpected_keys)
+                    )
+                )
+            if len(missing_keys) > 0:
+                error_msgs.insert(
+                    0, 'Missing key(s) in state_dict: {}. '.format(', '.join('"{}"'.format(k) for k in missing_keys))
+                )
+
+        if len(error_msgs) > 0:
+            raise RuntimeError(
+                'Error(s) in loading state_dict for {}:\n\t{}'.format(self.__class__.__name__, "\n\t".join(error_msgs))
+            )
+        return _IncompatibleKeys(missing_keys, unexpected_keys)